desnaking working dust clouds
I'm torn about the asynchronisity.
On one hand it helps see that it is an effect and not dangerous
On the other it wouldn't be as nervous at slow speeds.

compromise would be to adjust to game speed but stay independent
for the nice effect when worm dies or level is done.

The best part of trails is that you can see when a worm is growing (better).